==> Predicting the NBA MVP/3. Data_analysis.ipynb <==
# Data Analysis and Machine Learning [3/3]

import pandas as pd 
import numpy as np

stats = pd.read_csv("player_mvp_stats.csv", index_col=0)
stats

# we are gonna use the stats to predic thte votes everyone would get in the mvp voting

## 1. Looking at the Data / Cleaning

# checking null values 
pd.isnull(stats).sum()

stats[pd.isnull(stats['3P%'])][['Player', '3PA']]

stats[pd.isnull(stats['FT%'])][['Player', 'FTA']]

# so we can replace the missing percentages with 0
# technically not correct but we are doing so for the moment
stats = stats.fillna(0)

## 2. Training a model

stats.columns

#numeric columns
predictors = ['Age', 'G', 'GS', 'MP', 'FG', 'FGA', 'FG%', '3P',
              '3PA', '3P%', '2P', '2PA', '2P%', 'eFG%', 'FT', 'FTA', 'FT%', 'ORB',
              'DRB', 'TRB', 'AST', 'STL', 'BLK', 'TOV', 'PF', 'PTS', 'Year','W', 'L', 'W/L%', 'GB', 'PS/G', 'PA/G', 'SRS']

# we exclude ['Pts Won', 'Pts Max', 'Share'] because it's basically what we want to predict

train = stats[stats['Year']<2021]
test = stats[stats['Year']==2021]

from sklearn.linear_model import Ridge
# ridge is designed by Tikhonov to prevent overfitting
reg = Ridge(alpha=0.1) # alpha controls the shrinking of coefficients to prevent overfitting

reg.fit(train[predictors], train['Share'])

predictions = reg.predict(test[predictors])

predictions = pd.DataFrame(predictions, columns=['predictions'], index=test.index)

predictions

combination = pd.concat([test[['Player', 'Share']], predictions], axis=1)

combination #for every player has a share which is a percentage of the mvp votes

combination.sort_values('Share', ascending=False).head(10)

## 3. Identifying an error metric

from sklearn.metrics import mean_squared_error

mean_squared_error(combination['Share'], combination['predictions'])

combination['Share'].value_counts() #majority is zero so mse is not super insightful
# we just care about the top5 for example

#introduce rank 
combination = combination.sort_values('Share', ascending=False)
combination['Rk'] = list(range(1, combination.shape[0]+1))

combination.head(10)

# same for our predictions
combination = combination.sort_values('predictions', ascending=False)
combination['Pred_Rk'] = list(range(1, combination.shape[0]+1))

combination.head(10)

# so question is : Out of the top5 how many correct predictions did we get
# let's create one metric
# heavy penalties for deviations
combination.sort_values('Share', ascending=False).head(10)

def find_ap(combination):
    actual = combination.sort_values('Share', ascending=False).head(5) # get top5
    predicted = combination.sort_values('predictions', ascending=False)
    ps = []
    found = 0 
    seen = 1
    
    for index, row in predicted.iterrows():
        if row['Player'] in actual['Player'].values:
            found += 1
            ps.append(found/seen)
        seen += 1 
    return sum(ps) / len(ps)

find_ap(combination) # high values are better

## 4. Implementing Backtesting to predict each year

years = list(range(1991, 2022))

# this makes our metric more robust, fights overfitting
aps = []
all_predictions = []

for year in years[5:]:
    train = stats[stats['Year'] < year]
    test = stats[stats['Year'] == year]
    
    reg.fit(train[predictors], train['Share'])
    predictions = reg.predict(test[predictors])
    predictions = pd.DataFrame(predictions, columns=['predictions'], index=test.index)
    combination = pd.concat([test[['Player', 'Share']], predictions], axis=1)
    
    all_predictions.append(combination)
    aps.append(find_ap(combination))

#mean average precision
sum(aps)/len(aps)

def add_ranks(combination):
    combination = combination.sort_values('Share', ascending=False)
    combination['Rk'] = list(range(1, combination.shape[0]+1))

    combination = combination.sort_values('predictions', ascending=False)
    combination['Pred_Rk'] = list(range(1, combination.shape[0]+1))
    
    combination['Diff'] = combination['Rk'] - combination['Pred_Rk']
    return combination

ranking = add_ranks(all_predictions[1])
ranking[ranking['Rk']<5].sort_values('Diff', ascending=False)

#so let's make a function that does the same as before
def backtest(stats, model, year, predictors):
    aps = []
    all_predictions = []

    for year in years[5:]:
        train = stats[stats['Year'] < year]
        test = stats[stats['Year'] == year]

        model.fit(train[predictors], train['Share'])
        predictions = model.predict(test[predictors])
        predictions = pd.DataFrame(predictions, columns=['predictions'], index=test.index)
        combination = pd.concat([test[['Player', 'Share']], predictions], axis=1)
        combination = add_ranks(combination)
        all_predictions.append(combination)
        aps.append(find_ap(combination))    
    return sum(aps)/len(aps), aps, pd.concat(all_predictions)

mean_ap, aps, all_predictions = backtest(stats, reg, years[5:], predictors)
mean_ap

## 5. Diagnosing model performance

all_predictions[all_predictions['Rk']<=5].sort_values('Diff').head(10)  # what is the differnece with other mvp, why would he be missranked

# look at the coefficients of the regression

pd.concat([pd.Series(reg.coef_), pd.Series(predictors)], axis=1).sort_values(0, ascending=False)

## 6. Adding more predictors

stat_ratios = stats[['PTS', 'AST', 'STL', 'BLK', '3P', 'Year']].groupby('Year').apply(lambda x : x/x.mean())

stat_ratios

stats[['PTS_R', 'AST_R', 'STL_R', 'BLK_R', '3P_R']] = stat_ratios[['PTS','AST','STL','BLK','3P']]
stats.head()

predictors += ['PTS_R', 'AST_R', 'STL_R', 'BLK_R', '3P_R']

mean_ap, aps, all_predictions = backtest(stats, reg, years[5:], predictors)

mean_ap # there was a slight improvement

stats['NPos']=stats['Pos'].astype('category').cat.codes

stats['NTm']=stats['Tm'].astype('category').cat.codes

stats['NTm'].value_counts()

## 7. Using a Random Forest Model

from sklearn.ensemble import RandomForestRegressor

rf = RandomForestRegressor(n_estimators=50, random_state=42, min_samples_split=5)

mean_ap, aps, all_predictions = backtest(stats, rf, years[28:], predictors)

mean_ap

#compare to previous regressor 
mean_ap, aps, all_predictions = backtest(stats, reg, years[28:], predictors)
mean_ap

## 8. Discussion
